requirements.txt
asgiref==3.5.2
attrs==21.2.0
autopep8==1.6.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2
django-cors-headers==3.13.0
django-non-dark-admin==2.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
idna==2.10
iniconfig==1.1.1
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
Pillow==9.2.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.9.1
PyJWT==2.4.0
pyparsing==2.4.7
pytz==2022.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.3
tzdata==2022.1
urllib3==1.26.6
wrapt==1.12.1
